* avoid signed issues
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@20210 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/bzfs/RecordReplay.cxx
@@ -1381,7 +1381,7 @@ static bool setVariables(void *data)
   std::string key, value;
 
   data = nboUnpackUInt16(data, numVars);
-  for (int i = 0; i < numVars; i++) {
+  for (uint16_t i = 0; i < numVars; i++) {
     data = nboUnpackStdString(data, key);
     data = nboUnpackStdString(data, value);
     if (key == "poll") {
